collada options are now global and not on a per-feature-basis anymore
