+++ impexp-kml-collada-plugin/src/main/java/org/citydb/modules/kml/database/SolitaryVegetationObject.java
@@ -80,10 +80,6 @@ public class SolitaryVegetationObject extends KmlGenericObject{
 		return config.getKmlExportConfig().getVegetationDisplayForms();
 	}
 
-	public ColladaOptions getColladaOptions() {
-		return config.getKmlExportConfig().getVegetationColladaOptions();
-	}
-
 	public Balloon getBalloonSettings() {
 		return config.getKmlExportConfig().getVegetationBalloon();
 	}
@@ -195,10 +191,12 @@ public class SolitaryVegetationObject extends KmlGenericObject{
 					break;
 
 				case DisplayForm.COLLADA:
+					ColladaOptions colladaOptions = config.getKmlExportConfig().getColladaOptions();
+
 					String currentgmlId = getGmlId();
 					setGmlId(work.getGmlId());
 					setId(work.getId());
-					fillGenericObjectForCollada(rs, config.getKmlExportConfig().getVegetationColladaOptions().isGenerateTextureAtlases());
+					fillGenericObjectForCollada(rs, colladaOptions.isGenerateTextureAtlases());
 
 					if (currentgmlId != null && !currentgmlId.equals(work.getGmlId()) && getGeometryAmount() > GEOMETRY_AMOUNT_WARNING)
 						log.info("Object " + work.getGmlId() + " has more than " + GEOMETRY_AMOUNT_WARNING + " geometries. This may take a while to process...");
@@ -210,7 +208,6 @@ public class SolitaryVegetationObject extends KmlGenericObject{
 					}
 					setZOffset(zOffset);
 
-					ColladaOptions colladaOptions = getColladaOptions();
 					setIgnoreSurfaceOrientation(colladaOptions.isIgnoreSurfaceOrientation());
 					try {
 						if (work.getDisplayForm().isHighlightingEnabled())
